+++ examples/valkey/delete_instance.py
@@ -1,0 +1,13 @@
+import os
+
+from stackit.core.configuration import Configuration
+from stackit.valkey.api.default_api import DefaultApi
+
+project_id = os.getenv("PROJECT_ID")
+instance_id = "INSTANCE_ID"
+
+config = Configuration()
+client = DefaultApi(config)
+
+# Delete an instance
+client.delete_instance(project_id=project_id, instance_id=instance_id)